requirements.txt
aws-cdk-lib==2.85.0
constructs>=10.0.0,<11.0.0
sagemaker==2.152.0
